refactor: move components from layout folder to components folder
