+++ frontend/src/ui/components/Footer/Footer.scss
