+++ frontend/src/ui/components/Footer/Footer.tsx
