+++ frontend/src/ui/components/LoadingView/LoadingView.tsx
@@ -1,7 +1,7 @@
 import { Spinner } from '@blueprintjs/core';
 import React from 'react';
 
-export const GenericLoadingView: React.FC = () => (
+export const LoadingView: React.FC = () => (
   <div className="d-flex align-items-center justify-content-center h-100">
     <Spinner />
   </div>
