+++ frontend/src/ui/components/Search/Filters/Filters.scss
