+++ frontend/src/ui/components/Search/Filters/res/cuisines.ts
