+++ frontend/src/ui/components/Search/Filters/res/diets.ts
