+++ frontend/src/ui/components/Search/Filters/res/intolerances.ts
