+++ frontend/src/ui/components/Search/Filters/res/mealTypes.ts
